% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_hclust.R
\name{plot_missing_hclust}
\alias{plot_missing_hclust}
\title{Plot Missing Correlation Clustering}
\usage{
plot_missing_hclust(data, distance = "binary", ...)
}
\arguments{
\item{data}{a XXX}

\item{...}{any arguments to stats::hclust}
}
\value{
XXX
}
\description{
XXXX
}
\details{
XXX
}
\examples{
\dontrun{

}
}
\author{
Matt Simmons matthew.joseph.simmons@gmail.com
}
